server/requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu
transformers
pandas
xlrd
pdfminer.six
